Consolidated server files
